+++ src/lib/component/Linkbar.svelte
@@ -1,25 +1,31 @@
 <div class="flex-1 flex items-center flex-wrap gap-2 mt-8">
     <a
         href={`/`}
-        class="px-4 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
+        class="px-2 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
     >
         Home
     </a>
     <a
+        href={`/changelog`}
+        class="px-2 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
+    >
+        Changelog
+    </a>
+    <a
         href={`/resources`}
-        class="px-4 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
+        class="px-2 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
     >
         Resources
     </a>
     <a
         href={`/guides`}
-        class="px-4 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
+        class="px-2 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
     >
         Guides
     </a>
     <a
         href={`/disclaimer`}
-        class="px-4 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
+        class="px-2 py-2 bg-herobuttons hover:bg-herobuttonshover text-buttons font-extralight rounded-xl"
     >
         Disclaimer
     </a>